Changed the directory structure to adhere to the packaging guidelines
