+++ default_archs/TCN.py
@@ -88,7 +88,7 @@ import torch.nn as nn
 import numpy as np
 from torch.nn.utils import weight_norm
 
-from helpers.logger import get_logger
+from src.helpers.logger import get_logger
 logger = get_logger()
 
 available_tasks = ('regression', 'classification', 'forecasting', 'vl_regression')
