+++ environment.yml
@@ -1,233 +1,14 @@
 name: knowit_env
 channels:
-  - anaconda
-  - nvidia
   - pytorch
+  - nvidia
   - conda-forge
   - defaults
 dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - alabaster=0.7.12=pyhd3eb1b0_0
-  - appdirs=1.4.4=pyhd3eb1b0_0
-  - astroid=2.14.2=py38h06a4308_0
-  - attrs=23.1.0=py38h06a4308_0
-  - babel=2.11.0=py38h06a4308_0
-  - blas=1.0=mkl
-  - boost-cpp=1.74.0=h9359b55_0
-  - bottleneck=1.3.7=py38ha9d4c09_0
-  - brotli-python=1.0.9=py38h6a678d5_8
-  - brotlipy=0.7.0=py38h27cfd23_1003
-  - bzip2=1.0.8=h7b6447c_0
-  - ca-certificates=2025.1.31=hbcca054_0
-  - captum=0.6.0=0
-  - certifi=2024.2.2=pyhd8ed1ab_0
-  - cffi=1.15.1=py38h5eee18b_3
-  - cgal-cpp=5.0.3=h764656a_1
-  - charset-normalizer=2.0.4=pyhd3eb1b0_0
-  - click=8.1.7=py38h06a4308_0
-  - colorama=0.4.6=py38h06a4308_0
-  - configparser=5.0.2=pyhd3eb1b0_0
-  - cramjam=2.7.0=py38ha89cbab_0
-  - cryptography=41.0.3=py38hdda0065_0
-  - cuda-cudart=11.8.89=0
-  - cuda-cupti=11.8.87=0
-  - cuda-libraries=11.8.0=0
-  - cuda-nvrtc=11.8.89=0
-  - cuda-nvtx=11.8.86=0
-  - cuda-runtime=11.8.0=0
-  - cudatoolkit=11.8.0=h6a678d5_0
-  - cycler=0.11.0=pyhd3eb1b0_0
-  - dill=0.3.7=py38h06a4308_0
-  - docker-pycreds=0.4.0=pyhd3eb1b0_0
-  - docutils=0.18.1=py38h06a4308_3
-  - eigen=3.3.7=hd09550d_1
-  - exceptiongroup=1.2.0=py38h06a4308_0
-  - fastparquet=2024.2.0=py38ha9d4c09_0
-  - ffmpeg=4.2.2=h20bf706_0
-  - filelock=3.13.1=py38h06a4308_0
-  - freetype=2.12.1=h4a9f257_0
-  - giflib=5.2.1=h5eee18b_3
-  - gitdb=4.0.7=pyhd3eb1b0_0
-  - gitpython=3.1.37=py38h06a4308_0
-  - gmp=6.2.1=h295c915_3
-  - gmpy2=2.1.2=py38heeb90bb_0
-  - gnutls=3.6.15=he1e5248_0
-  - icu=67.1=he1b5a44_0
-  - idna=3.4=py38h06a4308_0
-  - imagesize=1.4.1=py38h06a4308_0
-  - importlib-metadata=6.0.0=py38h06a4308_0
-  - importlib_metadata=6.0.0=hd3eb1b0_0
-  - iniconfig=1.1.1=pyhd3eb1b0_0
-  - intel-openmp=2021.4.0=h06a4308_3561
-  - isort=5.9.3=pyhd3eb1b0_0
-  - jinja2=3.1.2=py38h06a4308_0
-  - joblib=1.2.0=py38h06a4308_0
-  - jpeg=9e=h5eee18b_1
-  - kiwisolver=1.4.4=py38h6a678d5_0
-  - lame=3.100=h7b6447c_0
-  - lazy-object-proxy=1.6.0=py38h27cfd23_0
-  - lcms2=2.12=h3be6417_0
-  - ld_impl_linux-64=2.38=h1181459_1
-  - libblas=3.9.0=12_linux64_mkl
-  - libcblas=3.9.0=12_linux64_mkl
-  - libcublas=11.11.3.6=0
-  - libcufft=10.9.0.58=0
-  - libcufile=1.7.2.10=0
-  - libcurand=10.3.3.141=0
-  - libcusolver=11.4.1.48=0
-  - libcusparse=11.7.5.86=0
-  - libffi=3.4.4=h6a678d5_0
-  - libgcc=14.2.0=h77fa898_1
-  - libgcc-ng=14.2.0=h69a702a_1
-  - libgfortran-ng=7.5.0=ha8ba4b0_17
-  - libgfortran4=7.5.0=ha8ba4b0_17
-  - libgomp=14.2.0=h77fa898_1
-  - libidn2=2.3.4=h5eee18b_0
-  - libjpeg-turbo=2.0.0=h9bf148f_0
-  - liblapack=3.9.0=12_linux64_mkl
-  - libnpp=11.8.0.86=0
-  - libnvjpeg=11.9.0.86=0
-  - libopus=1.3.1=h7b6447c_0
-  - libpng=1.6.39=h5eee18b_0
-  - libprotobuf=3.20.3=he621ea3_0
-  - libstdcxx=14.2.0=hc0a3c3a_1
-  - libstdcxx-ng=14.2.0=h4852527_1
-  - libtasn1=4.19.0=h5eee18b_0
-  - libtiff=4.2.0=h85742a9_0
-  - libunistring=0.9.10=h27cfd23_0
-  - libuv=1.44.2=h5eee18b_0
-  - libvpx=1.7.0=h439df22_0
-  - libwebp=1.3.2=h11a3e52_0
-  - libwebp-base=1.3.2=h5eee18b_0
-  - lightning-utilities=0.9.0=py38h06a4308_0
-  - llvm-openmp=14.0.6=h9e868ea_0
-  - lz4-c=1.9.4=h6a678d5_0
-  - markupsafe=2.1.1=py38h7f8727e_0
-  - matplotlib=3.2.2=1
-  - matplotlib-base=3.2.2=py38h5d868c9_1
-  - mccabe=0.7.0=pyhd3eb1b0_0
-  - mkl=2021.4.0=h06a4308_640
-  - mkl-service=2.4.0=py38h7f8727e_0
-  - mkl_fft=1.3.1=py38hd3c417c_0
-  - mkl_random=1.2.2=py38h51133e4_0
-  - more-itertools=8.12.0=pyhd3eb1b0_0
-  - mpc=1.1.0=h10f8cd9_1
-  - mpfr=4.0.2=hb69a4c5_1
-  - mpmath=1.3.0=py38h06a4308_0
-  - ncurses=6.4=h6a678d5_0
-  - nettle=3.7.3=hbbd107a_1
-  - networkx=3.1=py38h06a4308_0
-  - nlopt=2.7.0=py38h6599729_1
-  - numexpr=2.8.4=py38he184ba9_0
-  - numpy=1.24.3=py38h14f4228_0
-  - numpy-base=1.24.3=py38h31eccc5_0
-  - openh264=2.1.1=h4ff587b_0
-  - openjpeg=2.4.0=h3ad879b_0
-  - openssl=3.4.1=h7b32b05_0
-  - packaging=23.1=py38h06a4308_0
-  - pandas=1.5.3=py38hdc8b05c_1
-  - pathtools=0.1.2=pyhd3eb1b0_1
-  - pillow=10.0.1=py38ha6cbd5a_0
-  - pip=23.3=py38h06a4308_0
-  - platformdirs=3.10.0=py38h06a4308_0
-  - pluggy=1.0.0=py38h06a4308_1
-  - pockets=0.9.1=py_0
-  - promise=2.3=py38h06a4308_0
-  - protobuf=3.20.3=py38h6a678d5_0
-  - psutil=5.9.0=py38h5eee18b_0
-  - py=1.11.0=pyhd3eb1b0_0
-  - pycparser=2.21=pyhd3eb1b0_0
-  - pygments=2.15.1=py38h06a4308_1
-  - pylint=2.16.2=py38h06a4308_0
-  - pynput=1.7.3=py38h578d9bd_0
-  - pyopenssl=23.2.0=py38h06a4308_0
-  - pyparsing=3.0.9=py38h06a4308_0
-  - pysocks=1.7.1=py38h06a4308_0
-  - pytest=7.4.0=py38h06a4308_0
-  - python=3.8.18=h955ad1f_0
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - python-xlib=0.28=pyh9f0ad1d_0
-  - python_abi=3.8=2_cp38
-  - pytorch=2.1.0=py3.8_cuda11.8_cudnn8.7.0_0
-  - pytorch-cuda=11.8=h7e8668a_5
-  - pytorch-lightning=2.0.3=py38h06a4308_0
-  - pytorch-mutex=1.0=cuda
-  - pytz=2023.3.post1=py38h06a4308_0
-  - pyyaml=6.0.1=py38h5eee18b_0
-  - readline=8.2=h5eee18b_0
-  - requests=2.31.0=py38h06a4308_0
-  - scikit-geometry=0.1.2=py38he18db2a_4
-  - scikit-learn=1.3.0=py38h1128e8f_0
-  - scipy=1.5.3=py38h828c644_0
-  - sentry-sdk=1.9.0=py38h06a4308_0
-  - setproctitle=1.2.2=py38h27cfd23_1004
-  - setuptools=68.0.0=py38h06a4308_0
-  - setuptools-lint=0.6.0=pyh9f0ad1d_0
-  - shortuuid=1.0.8=py38h578d9bd_0
-  - six=1.16.0=pyhd3eb1b0_1
-  - smmap=4.0.0=pyhd3eb1b0_0
-  - snowballstemmer=2.2.0=pyhd3eb1b0_0
-  - sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
-  - sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
-  - sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
-  - sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
-  - sphinxcontrib-napoleon=0.7=py_0
-  - sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
-  - sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
-  - sqlite=3.41.2=h5eee18b_0
-  - sympy=1.13.3=py38h06a4308_0
-  - threadpoolctl=2.2.0=pyh0d69192_0
-  - tk=8.6.12=h1ccaba5_0
-  - toml=0.10.2=pyhd3eb1b0_0
-  - tomli=2.0.1=py38h06a4308_0
-  - tomlkit=0.13.2=py38h06a4308_0
-  - torchaudio=2.1.0=py38_cu118
-  - torchtriton=2.1.0=py38
-  - torchvision=0.16.0=py38_cu118
-  - tornado=6.3.3=py38h5eee18b_0
-  - tqdm=4.66.4=py38h2f386ee_0
-  - typing-extensions=4.7.1=py38h06a4308_0
-  - typing_extensions=4.7.1=py38h06a4308_0
-  - urllib3=1.26.18=py38h06a4308_0
-  - wandb=0.16.6=pyhd8ed1ab_1
-  - wheel=0.41.2=py38h06a4308_0
-  - wrapt=1.11.2=py38h1e0a361_1
-  - x264=1!157.20191217=h7b6447c_0
-  - xz=5.4.2=h5eee18b_0
-  - yaml=0.2.5=h7b6447c_0
-  - zipp=3.11.0=py38h06a4308_0
-  - zlib=1.2.13=h5eee18b_0
-  - zstd=1.4.9=haebb681_0
+  - python=3.10
+  - pip
+  - boost-cpp
+  - cgal-cpp
+  - cudatoolkit=11.8
   - pip:
-    - accessible-pygments==0.0.4
-    - beautifulsoup4==4.12.3
-    - cloudpickle==2.0.0
-    - configspace==0.5.0
-    - cython==0.29.28
-    - dask==2022.5.0
-    - distributed==2022.5.0
-    - emcee==3.1.1
-    - fsspec==2022.3.0
-    - heapdict==1.0.1
-    - locket==1.0.0
-    - markdown-it-py==3.0.0
-    - mdit-py-plugins==0.4.1
-    - mdurl==0.1.2
-    - msgpack==1.0.3
-    - myst-parser==3.0.1
-    - partd==1.2.0
-    - pyarrow==17.0.0
-    - pydata-sphinx-theme==0.14.4
-    - pynisher==0.6.4
-    - pyrfr==0.8.2
-    - smac==1.3.1
-    - sortedcontainers==2.4.0
-    - soupsieve==2.5
-    - sphinx==7.1.2
-    - sphinx-autoapi==3.1.1
-    - tblib==1.7.0
-    - toolz==0.11.2
-    - torchmetrics==1.5.2
-    - zict==2.2.0
-prefix: /home/tian/miniconda3/envs/knowit_env
+      - -e .[torch,dev,extra-data]
