--- setup/__init__.py
@@ -1,7 +1,0 @@
-"""KnowIt setup functions."""
-
-from setup import import_custom_arch, select_interpretation_points, setup_action_args, setup_weighted_cross_entropy
-
-__all__ = [
-
-]
