+++ src/data/data_splitting.py
@@ -79,8 +79,8 @@ from numpy import (array, random, argwhere, isnan,
                    argsort, arange, unique, append, floor, full)
 
 # internal imports
-from data.base_dataset import DataExtractor
-from helpers.logger import get_logger
+from src.data.base_dataset import DataExtractor
+from src.helpers.logger import get_logger
 logger = get_logger()
