+++ src/data/prepared_dataset.py
@@ -185,10 +185,10 @@ from torch import from_numpy, is_tensor, Tensor, unsqueeze
 from torch import zeros as zeros_tensor
 
 # internal imports
-from data.base_dataset import BaseDataset
-from data.data_splitting import DataSplitter
-from data.data_scaling import DataScaler
-from helpers.logger import get_logger
+from src.data.base_dataset import BaseDataset
+from src.data.data_splitting import DataSplitter
+from src.data.data_scaling import DataScaler
+from src.helpers.logger import get_logger
 
 logger = get_logger()
