+++ src/helpers/file_dir_procs.py
@@ -13,11 +13,10 @@ import _pickle as cPickle
 import gzip
 import lzma
 import shutil
-import csv
 import pandas
 
 # internal imports
-from helpers.logger import get_logger
+from src.helpers.logger import get_logger
 
 logger = get_logger()
