+++ src/setup/__init__.py
@@ -1,0 +1,8 @@
+"""KnowIt setup functions."""
+
+from src.setup import setup_action_args, select_interpretation_points, import_custom_arch, \
+    setup_weighted_cross_entropy
+
+__all__ = [
+
+]
