+++ src/setup/setup_action_args.py
@@ -23,7 +23,7 @@ __description__ = 'Checks, filters, and adjusts user arguments for various actio
 # 'default' arguments will be used in place of optional if not provided
 
 # internal imports
-from helpers.logger import get_logger
+from src.helpers.logger import get_logger
 logger = get_logger()
 
 arg_dict = {'data_import':  {'required': ('raw_data',),
