+++ src/setup/setup_scripts.md
