+++ src/trainer/trainer_framework.md
